+++ CHANGELOG.md
@@ -14,8 +14,8 @@ All notable changes to this project will be documented in this file.
 
 - Added a lint rule to verify field `private_key` for the `snowflake_streaming` output is in PEM format. (@rockwotj)
 - New `mongodb_cdc` input for change data capture (CDC) over MongoDB collections. (@rockwotj)
-- Field `is_end_offset` added to the `redpanda_migrator_offsets` output. (@mihaitodor)
-- Metadata field `kafka_is_end_offset` added to the `redpanda_migrator_offsets` input. (@mihaitodor)
+- Field `is_high_watermark` added to the `redpanda_migrator_offsets` output. (@mihaitodor)
+- Metadata field `kafka_is_high_watermark` added to the `redpanda_migrator_offsets` input. (@mihaitodor)
 
 ### Changed